Add parallel by id gene queries
As well as get_gene_info add get_gene_info_by_id, as well as
get_tx_for_gene add get_gene_info_by_id, the original queries are via
hgnc gene symbol, the new are by hgnc stable id. These new functions
rely on the presence of the hgnc_id column, which is only present in
vvta 0.6+, however there is no way to specify this without causing
failure to run at all on old databases. This would be harmful at this
point, so we just have to accept that using the new queries will case
a crash on old databases for now.