Add new materialised view based filtered queries
Add an upgraded set of vvta specific materialised view based queries to
be activated only on vvta views>0.7, this adds another check to startup
so will be removed in favor of requiring vvta later. These materialised
views contain data that has been filtered both to limit it to non
"removed" transcripts and to pre-run the joins and other such operations.