Store the relative accession for transcripts
This is the origin accession for variants mapped from the genome but may
also be user specified, though the parser is not updated to do this yet.
This paired accesion is then used in preference for intron exon boundary
validation and as a default for remapping against the genome in the easy
variant mapper.

Although this change is positive in most cases it does shift the
behaviour of mappings enough that VV changes are needed to compensate in
the cases when we *do* intend to force a remap onto the main genomic
chromosomes. Since the biggest gain for this change is WRT transcripts
with varying intron/exon boundaries this patch is not UTA compatible or
relevant.